LaptopsDirect/requirements.txt
scrapy-fake-useragent==1.4.0